Restrict translating and setting label values to only handle numbers
For strict type checking, the translate function should take only
numeric argument, but meanwhile, combined label value was passed as
a string, which is not ideal at all.

This change effectively removes the ability to translate combined
labels. It could be brought back later but using a solution that is
strictly type checked.