Use custom templates for label elements instead of HTML injection
Yay, no more injecting raw HTML!
But the slider is now broken :(

This is because updates to the label values are not translated to DOM
updates right away and the subsequent code calculating positions is
operating on previous values.

To fix this, the slider update code has to be rewritten to account for
that. Ideally, the code flow, starting from some event handler, should be:
 1. Run the code until label updates.
 2. Update label values
 3. Later, when DOM changes are finished, reposition the elements.
 4. Run rest of the code in the event handler.