Remove id option
It should not be necessary, as the users should be able to bind their own
functions which capture the desired behaviour.

It also makes all custom functions passed in optios consistent with
each other, as the id was only supplied to the translate() and getLegend()
functions but not to the other ones.