Make ng-bootstrap a truly optional dependency
ng-bootstrap is a peer dependency, so NPM will allow installation
without it. However, it would still result in errors in runtime
due to the way the compiled code references ng-bootstrap module.

This fixes the errors, making ng-bootstrap truly optional.
There are two side effects of this change:

 - using ng5-slider without ng-bootstrap will result in warning
   messages during compilation about missing dependency to ng-bootstrap,
   but these can be safely ignored

 - ahead-of-time (AOT) compilation is no longer possible, in the app
   using ng5-slider, even if ng-bootstrap is available

Unavailable AOT compilation is sad news, but it just seems that
conditional logic defeats the AOT compiler, and there are multiple
issues reported regarding this. Hopefully, this will be fixed at some
point.