Implement tooltips using user-specified templates
Tooltips are now implemented through user-specified template
which gets passed the context information (tooltip, placement, context)
and is responsible for rendering the tooltip element with whatever user
library.

If the template is not provided, the tooltip falls back to using
standard title HTML attribute. Placement is ignored in that case.