Apply changes in inputs in correct order (issue #56)
This fixes a bug that was introduced by recent addition of normalisation
and range checking. When changing both options and model values at the
same time, it becomes necessary to apply the options first, then the model
values, so that normalisation code will pick up the correct floor-ceil
range.

This directly fixes issue #56, but may also fix similar problems
described in issues #54 and #58.