Prevent unnecessary updates on internal slider changes
This fixes a regression introduced with recent addition of ngOnChanges.

If user interacted with the slider, and the slider updated its view,
and was pushing the change to value bindings, then ngOnChanges would be
triggered unnecessarily.

This fixes it so that ngOnChanges is only triggered on genuine
externally applied changes (coming from parent component), and internal
changes (coming from event handling), are ignored.