Generate API documentation in new way
Upgraded typedoc no longer allows to modify theme files quite so easily
as before, allowing the documentation to be part of the Angular demo
app.

Therefore, the generated documentation files are now left untouched,
generated using the default typedoc theme, and instead of embedding them
as part of the demo app, they will now be rendered in an iframe.