Fix pointer event not being cleaned up in Safari
Touch events that touch a target (the slider) and then trigger
scrolling, will not move the slider and also not trigger `pointerup`
for the original element if the touch was finished on a different
element. `pointercancel` is triggered when all pointer events are
finished (touch in our case), thus it can be used to clean up any
lingering pointer events.

Another problem when opening the context-menu - even in mouse context -
Safari won't trigger a `pointerup` event. Thus also have to listen to
`pointercancel` for mouse events.